refactor: 예외 클래스 및 핸들러 패키지 위치 수정
- ChatRoomNotFoundException 및 UnauthorizedChatRoomAccessException 패키지 이동
- fittoring.mentoring.business.exception → fittoring.application.exception으로 수정
- ChatRoomService, ChatMessageService 등에서 예외 클래스 참조 변경
- GlobalExceptionHandler의 예외 처리 메서드에서 변경된 경로 업데이트
- 관련 테스트 코드의 import 경로 수정 및 반영
