+++ backend/fittoring/src/main/java/fittoring/application/chat/service/ChatMessageService.java
@@ -7,10 +7,10 @@ import fittoring.application.chat.repository.ChatMessageRepository;
 import fittoring.application.chat.repository.ChatRoomRepository;
 import fittoring.application.chat.service.dto.ChatMessagePaginationResultDto;
 import fittoring.application.exception.BusinessErrorMessage;
+import fittoring.application.exception.ChatRoomNotFoundException;
+import fittoring.application.exception.UnauthorizedChatRoomAccessException;
 import fittoring.domain.model.ChatMessage;
 import fittoring.domain.model.ChatRoom;
-import fittoring.mentoring.business.exception.ChatRoomNotFoundException;
-import fittoring.mentoring.business.exception.UnauthorizedChatRoomAccessException;
 import fittoring.util.Cursor;
 import fittoring.util.CursorCodec;
 import java.util.List;
