+++ backend/fittoring/src/main/java/fittoring/application/chat/service/ChatRoomService.java
@@ -5,9 +5,11 @@ import fittoring.application.chat.service.dto.ChatRoomCreatedInfo;
 import fittoring.application.chat.service.dto.ChatRoomInfoDto;
 import fittoring.application.exception.BusinessErrorMessage;
 import fittoring.application.exception.ChatRoomAlreadyExistsException;
+import fittoring.application.exception.ChatRoomNotFoundException;
 import fittoring.application.exception.MemberNotFoundException;
 import fittoring.application.exception.MentoringNotFoundException;
 import fittoring.application.exception.ReservationNotFoundException;
+import fittoring.application.exception.UnauthorizedChatRoomAccessException;
 import fittoring.application.member.repository.MemberRepository;
 import fittoring.application.mentoring.repository.MentoringRepository;
 import fittoring.application.reservation.repository.ReservationRepository;
@@ -80,14 +82,13 @@ public class ChatRoomService {
     private ChatRoom getChatRoom(Long chatroomId) {
         return chatRoomRepository.findById(chatroomId)
                 .orElseThrow(
-                        () -> new fittoring.mentoring.business.exception.ChatRoomNotFoundException(
-                                BusinessErrorMessage.CHAT_ROOM_NOT_FOUND.getMessage())
+                        () -> new ChatRoomNotFoundException(BusinessErrorMessage.CHAT_ROOM_NOT_FOUND.getMessage())
                 );
     }
 
     private void validateParticipant(Long memberId, ChatRoom chatRoom) {
         if (chatRoom.isNonParticipant(memberId)) {
-            throw new fittoring.mentoring.business.exception.UnauthorizedChatRoomAccessException(
+            throw new UnauthorizedChatRoomAccessException(
                     BusinessErrorMessage.UNAUTHORIZED_CHAT_ROOM_ACCESS.getMessage()
             );
         }
@@ -103,7 +104,7 @@ public class ChatRoomService {
 
     private void validateReservationStatus(Reservation reservation) {
         if (!reservation.isAccessibleForChatRoom()) {
-            throw new fittoring.mentoring.business.exception.UnauthorizedChatRoomAccessException(
+            throw new UnauthorizedChatRoomAccessException(
                     BusinessErrorMessage.INVALID_STATUS_CHAT_ROOM_ACCESS.getMessage()
             );
         }
