+++ backend/fittoring/src/main/java/fittoring/application/exception/ChatRoomNotFoundException.java
@@ -1,4 +1,4 @@
-package fittoring.mentoring.business.exception;
+package fittoring.application.exception;
 
 public class ChatRoomNotFoundException extends RuntimeException {
