+++ backend/fittoring/src/main/java/fittoring/exception/GlobalExceptionHandler.java
@@ -5,6 +5,7 @@ import fittoring.application.exception.BusinessErrorMessage;
 import fittoring.application.exception.CategoryNotFoundException;
 import fittoring.application.exception.CertificateNotFoundException;
 import fittoring.application.exception.ChatRoomAlreadyExistsException;
+import fittoring.application.exception.ChatRoomNotFoundException;
 import fittoring.application.exception.DuplicateLoginIdException;
 import fittoring.application.exception.DuplicatePhoneException;
 import fittoring.application.exception.ForbiddenException;
@@ -26,12 +27,11 @@ import fittoring.application.exception.ReservationNotCompletedException;
 import fittoring.application.exception.ReservationNotFoundException;
 import fittoring.application.exception.ReviewAlreadyExistsException;
 import fittoring.application.exception.ReviewNotFoundException;
+import fittoring.application.exception.UnauthorizedChatRoomAccessException;
 import fittoring.application.exception.UnsupportedImageExtensionException;
 import fittoring.infrastructure.exception.S3UploadException;
 import fittoring.infrastructure.exception.SmsException;
 import fittoring.logging.dto.ErrorLog;
-import fittoring.mentoring.business.exception.ChatRoomNotFoundException;
-import fittoring.mentoring.business.exception.UnauthorizedChatRoomAccessException;
 import fittoring.util.ResponseDurationCalculator;
 import java.time.LocalDateTime;
 import lombok.RequiredArgsConstructor;
