+++ backend/fittoring/src/test/java/fittoring/application/auth/service/ChatMessageServiceTest.java
@@ -8,9 +8,9 @@ import fittoring.application.chat.presentation.dto.request.ChatMessageRequest;
 import fittoring.application.chat.repository.ChatRoomRepository;
 import fittoring.application.chat.service.ChatMessageService;
 import fittoring.application.exception.BusinessErrorMessage;
+import fittoring.application.exception.ChatRoomNotFoundException;
+import fittoring.application.exception.UnauthorizedChatRoomAccessException;
 import fittoring.domain.model.ChatRoom;
-import fittoring.mentoring.business.exception.ChatRoomNotFoundException;
-import fittoring.mentoring.business.exception.UnauthorizedChatRoomAccessException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
